+++ src/models/gamma/hmm_g4.jl
@@ -1,4 +1,4 @@
-mutable struct ZebrafishHMM <: HiddenMarkovModels.AbstractHMM
+mutable struct ZebrafishHMM_G4 <: HiddenMarkovModels.AbstractHMM
     #=
     states are stored in this order:
         1. Forward-Left (forward bouts, but last turn was left)
@@ -11,7 +11,7 @@ mutable struct ZebrafishHMM <: HiddenMarkovModels.AbstractHMM
     forw::Normal{Float64}
     turn::Gamma{Float64}
 
-    function ZebrafishHMM(
+    function ZebrafishHMM_G4(
         initial_probs::AbstractVector{<:Real},
         transition_matrix::AbstractMatrix{<:Real},
         forw::Normal{<:Real}, turn::Gamma{<:Real}
@@ -23,17 +23,17 @@ mutable struct ZebrafishHMM <: HiddenMarkovModels.AbstractHMM
 end
 
 # number of hidden states
-Base.length(hmm::ZebrafishHMM) = length(hmm.initial_probs)
+Base.length(hmm::ZebrafishHMM_G4) = length(hmm.initial_probs)
 
-function HiddenMarkovModels.transition_matrix(hmm::ZebrafishHMM)
+function HiddenMarkovModels.transition_matrix(hmm::ZebrafishHMM_G4)
     return hmm.transition_matrix
 end
 
-function HiddenMarkovModels.initial_distribution(hmm::ZebrafishHMM)
+function HiddenMarkovModels.initial_distribution(hmm::ZebrafishHMM_G4)
     return hmm.initial_probs
 end
 
-function HiddenMarkovModels.obs_distribution(hmm::ZebrafishHMM, i::Int)
+function HiddenMarkovModels.obs_distribution(hmm::ZebrafishHMM_G4, i::Int)
     if i ∈ (1, 2) # forward-left, forward-right
         dist = hmm.forw
     elseif i == 3 # left
@@ -44,20 +44,20 @@ function HiddenMarkovModels.obs_distribution(hmm::ZebrafishHMM, i::Int)
         throw(ArgumentError("State index must be 1, 2, 3, or 4; got $i"))
     end
 
-    return DistributionMissingWrapper(dist)
+    return dist
 end
 
-function normalize_initial_probs!(hmm::ZebrafishHMM)
+function normalize_initial_probs!(hmm::ZebrafishHMM_G4)
     hmm.initial_probs .= hmm.initial_probs ./ sum(hmm.initial_probs)
     return hmm.initial_probs
 end
 
-function normalize_transition_matrix!(hmm::ZebrafishHMM)
+function normalize_transition_matrix!(hmm::ZebrafishHMM_G4)
     # forbidden transitions
     hmm.transition_matrix[3,2] = 0 # forbid left -> forward-right
     hmm.transition_matrix[4,1] = 0 # forbid right -> forward-left
-    hmm.transition_matrix[1,2] = 0 # forbid forward-left -> forward-right
-    hmm.transition_matrix[2,1] = 0 # forbid forward-right -> forward-left
+    #hmm.transition_matrix[1,2] = 0 # forbid forward-left -> forward-right
+    #hmm.transition_matrix[2,1] = 0 # forbid forward-right -> forward-left
 
     # TODO: impose left / right symmetry ?
 
@@ -67,13 +67,13 @@ function normalize_transition_matrix!(hmm::ZebrafishHMM)
     return hmm.transition_matrix
 end
 
-function normalize_all!(hmm::ZebrafishHMM)
+function normalize_all!(hmm::ZebrafishHMM_G4)
     normalize_initial_probs!(hmm)
     normalize_transition_matrix!(hmm)
     return hmm
 end
 
-function StatsAPI.fit!(hmm::ZebrafishHMM, init_count, trans_count, obs_seq, state_marginals)
+function StatsAPI.fit!(hmm::ZebrafishHMM_G4, init_count, trans_count, obs_seq, state_marginals)
     #= Update initial state probabilities =#
     hmm.initial_probs .= init_count
     normalize_initial_probs!(hmm)
@@ -82,26 +82,10 @@ function StatsAPI.fit!(hmm::ZebrafishHMM, init_count, trans_count, obs_seq, stat
     hmm.transition_matrix .= trans_count
     normalize_transition_matrix!(hmm)
 
-    #= Update forward emission probabilities =#
-    forw_obs, forw_marginals = filter_obs(
-        !ismissing, obs_seq, state_marginals[1,:] + state_marginals[2,:]
-    )
-    forw = fit(typeof(hmm.forw), forw_obs, forw_marginals)
+    #= Update forward emission probabilities. =#
+    hmm.forw = fit_mle(typeof(hmm.forw), obs_seq, state_marginals[1,:] + state_marginals[2,:]; mu = 0.0)
 
-    # forward angles are always centered at 0, so we only use the fitted σ, discarding μ
-    hmm.forw = Normal(0, forw.σ)
-
-    #= Update left-right turn emission probabilities =#
-    # discard negative and missing entries, since they could not have been emitted by turning states
-    turn_obs, turn_marginals = filter_obs(
-        x -> !ismissing(x) && x > 0,
-        [-obs_seq; obs_seq],
-        [state_marginals[3,:]; state_marginals[4,:]]
-    )
-    hmm.turn = fit(typeof(hmm.turn), turn_obs, turn_marginals)
-end
-
-function filter_obs(cond, obs_seq::AbstractVector, state_marginals::AbstractVector)
-    _idx = findall(cond, obs_seq)
-    return map(identity, obs_seq[_idx]), map(identity, state_marginals[_idx])
+    #= Update left-right turn emission probabilities. =#
+    turn_marginals = ifelse.(obs_seq .< 0, state_marginals[3,:], state_marginals[4,:])
+    hmm.turn = fit_mle(typeof(hmm.turn), abs.(obs_seq), turn_marginals)
 end