+++ test/runtests.jl
@@ -1,2 +1,3 @@
 module aqua_tests include("aqua.jl") end
 module artifacts_tests include("artifacts.jl") end
+module truncated_normal_tests include("truncated_normal.jl") end